scm_document_manager/requirements.txt
# Core dependencies
streamlit==1.30.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google APIs
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==5.12.3
oauth2client==4.1.3

# Logging and monitoring
structlog==23.3.0

# Retry logic
tenacity==8.2.3

# Document processing (Phase 2)
# pdfplumber==0.10.3
# pytesseract==0.3.10
# openpyxl==3.1.2
# pandas==2.1.4

# AI/Vector DB (Phase 2)
# chromadb==0.4.22
# sentence-transformers==2.2.2
# google-generativeai==0.3.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
